website/services/warranty-analyzer/pyproject.toml
[project]
name = "warranty-analyzer"
version = "1.0.0"
description = "AI-powered warranty contract analysis service using Claude AI"
requires-python = ">=3.12"
authors = [
    {name = "Snapregister Team", email = "dev@snapregister.com"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "anthropic>=0.18.1",
    "PyPDF2>=3.0.1",
    "pdf2image>=1.17.0",
    "Pillow>=10.2.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.14",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
